added the rest of the structure in more organized manner
